finally done

+++ app/Http/Controllers/BlogController.php
@@ -3,37 +3,28 @@
 namespace App\Http\Controllers;
 
 use App\Models\Blog;
+use Facade\FlareClient\View;
 use Illuminate\Http\Request;
 
 class BlogController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    public function __construct()
+    {
+        // Info: Make sure the user is login.
+        $this->middleware(['auth'])->only([
+            'store',
+            'destroy',
+            'edit',
+            'update',
+        ]);
+    }
+
     public function index()
     {
         $blogs = Blog::latest()->paginate(10);
         return view('blogs.index', ['blogs' => $blogs]);
     }
 
-    // /**
-    //  * Show the form for creating a new resource.
-    //  *
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function create()
-    // {
-    //     //
-    // }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         $this->validate($request, ['body' => 'required']);
@@ -44,49 +35,35 @@ class BlogController extends Controller
         return back();
     }
 
-    // /**
-    //  * Display the specified resource.
-    //  *
-    //  * @param  \App\Models\Blog  $blog
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function show(Blog $blog)
-    // {
-    //     //
-    // }
+    public function show(Blog $blog)
+    {
+        return view('blogs.show', [
+            'blog' => $blog,
+        ]);
+    }
 
-    // /**
-    //  * Show the form for editing the specified resource.
-    //  *
-    //  * @param  \App\Models\Blog  $blog
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function edit(Blog $blog)
-    // {
-    //     //
-    // }
+    public function edit(Blog $blog)
+    {
+        return view('blogs.edit', [
+            'blog' => $blog,
+        ]);
+    }
 
-    // /**
-    //  * Update the specified resource in storage.
-    //  *
-    //  * @param  \Illuminate\Http\Request  $request
-    //  * @param  \App\Models\Blog  $blog
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function update(Request $request, Blog $blog)
-    // {
-    //     //
-    // }
+    // Info: Req params $id can be type cast to the model object or the id it self.
+    public function update(Request $request, $id)
+    {
+        $this->validate($request, ['body' => 'required']);
+        $targetBlog = Blog::find($id);
+        $this->authorize('update', $targetBlog);
+        $targetBlog->body = $request->body;
+        $targetBlog->save();
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Blog  $blog
-     * @return \Illuminate\Http\Response
-     */
+        return redirect(route('blogs'));
+    }
+
     public function destroy(Blog $blog)
     {
-        // Todo : Check isAdmin & authorization
+        $this->authorize('delete', $blog);
         $blog->delete();
         return back();
     }
